requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
elasticsearch==8.9.0
langchain==0.0.280
sentence-transformers==2.2.2
scikit-learn==1.3.0
pydantic==1.10.7
streamlit==1.25.0
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pinecone-client==6.1.0
apscheduler==3.10.1

